+++ src/app/application/mod.rs
@@ -1,0 +1,2 @@
+pub mod generate_handler;
+pub mod pull_templates_handler;