+++ src/cfg/domain/templates.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Deserialize, Serialize, Clone)]
+pub struct Templates {
+    pub repository: String,
+    pub branch: String,
+    pub remote: String,
+}