+++ src/core/port/mod.rs
@@ -1,0 +1,4 @@
+pub mod filesystem_port;
+pub mod logger_port;
+pub mod repository_port;
+pub mod templates_port;